+++ ContextHub.framework/Versions/A/Headers/CCHSubscriptionService.h
@@ -8,7 +8,6 @@
 
 #import <Foundation/Foundation.h>
 
-
 /**
  ContextHub Subscription error codes.
  */
@@ -38,7 +37,6 @@ extern NSString * const CCHBeaconUpdatedNotification;
  Posted to tagged beacon subscribers when a corresponding tagged beacon is deleted.
  */
 extern NSString * const CCHBeaconDeletedNotification;
-
 
 /**
  Posted to tagged geofence subscribers when a corresponding tagged geofence is created.
@@ -70,13 +68,47 @@ extern NSString * const CCHVaultItemUpdatedNotification;
  */
 extern NSString * const CCHVaultItemDeletedNotification;
 
+/**
+ Posted to tagged device subscribers when a corresponding tagged device is created.
+ */
+extern NSString * const CCHDeviceCreatedNotification;
+
+/**
+ Posted to tagged device subscribers when a corresponding tagged device is updated.
+ */
+extern NSString * const CCHDeviceUpdatedNotification;
+
+/**
+ Posted to tagged device subscribers when a corresponding tagged device is deleted.
+ */
+extern NSString * const CCHDeviceDeletedNotification;
+
+/**
+ Beacon option for adding and removing subscriptions.
+ */
+extern NSString * const CCHOptionBeacon;
+
+/**
+ Geofence option for adding and removing subscriptions.
+ */
+extern NSString * const CCHOptionGeofence;
+
+/**
+ Vault option for adding and removing subscriptions.
+ */
+extern NSString * const CCHOptionVault;
+
+/**
+ Device option for adding and removing subscriptions.
+ */
+extern NSString * const CCHOptionDevice;
 
 
 #define kSubscriptionErrorDomain @"com.contexthub.subscription"
 
 /**
- The subscription service is used tell ContextHub that you want to be notified when tagged elements are created, updated, and deleted on the server.  You must enable configure the push notifications if you want to receive updates from the server.
-When changes are detected, the subscription service will post notifications to the NSNotificationCenter.
+ The subscription service is used tell ContextHub that you want to be notified when tagged elements are created, updated, and deleted on the server.  You must enable push notifications if you want to receive updates from the server.
+When server changes are made, the device is notified using a background push notification.  The subscription service will post notifications to the NSNotificationCenter when changes are detected.
 
  ## Notifications
  
@@ -109,6 +141,15 @@ When changes are detected, the subscription service will post notifications to t
  ### CCHVaultItemDeletedNotification
  the notification object is an id of the vualt item that was deleted.  The userInfo object is not set.
 
+ ### CCHDeviceCreatedNotification
+ the notification object is an NSDictionary representation of the device that was created.  The userInfo object is not set.
+ 
+ ### CCHDeviceUpdatedNotification
+ the notification object is an NSDictionary of representation of the device that was updated.  The userInfo object is not set.
+ 
+ ### CCHDeviceDeletedNotification
+ the notification object is an id of the device that was deleted.  The userInfo object is not set.
+
  */
 @interface CCHSubscriptionService : NSObject
 
@@ -120,9 +161,9 @@ When changes are detected, the subscription service will post notifications to t
 /**
  Gets all subscriptions for the current device.
  @note Access individual subscriptions using "BeaconSubscription" and "GeofenceSubscription" keys
- @param completion executed when the request completes.  The block is passed an NSDictionary of subscriptions.  If an error occurs, the NSError will be passed to the block.
+ @param completionHandler executed when the request completes.  The block is passed an NSDictionary of subscriptions.  If an error occurs, the NSError will be passed to the block.
  */
-- (void)getSubscriptionsWithCompletion:(void(^)(NSDictionary *subscriptions, NSError *error))completion;
+- (void)getSubscriptionsWithCompletionHandler:(void(^)(NSDictionary *subscriptions, NSError *error))completionHandler;
 
 /**
  Subscribes the device to beacon change notifications for the specified tags.
@@ -152,5 +193,20 @@ When changes are detected, the subscription service will post notifications to t
  */
 - (void)removeGeofenceSubscriptionForTags:(NSArray *)tags completionHandler:(void(^)(NSError *error))completionHandler;
 
+/**
+ Subscribes the device to change notifications for the specified tags.
+ @note This will turn on background push notifications for all elements that have tags matching the tags array specified.  You must enable push notifications, enbale remote notifications and background fetch capabilites, and you must call application:didReceiveRemoteNotification:completionHandler: on CCHPush.
+ @param tags An NSArray of tags
+ @param options (optional) an NSArray of the elements that you want to subscribe to. (CCHOptionBeacon, CCHOptionGeofence, CCHOptionVault, CCHOptionDevice)
+ @param completionHandler (optional) Is executed when the request completes.  If an error occurs, the NSError will be passed to the block.
+ */
+- (void)addSubscriptionsForTags:(NSArray *)tags options:(NSArray *)options completionHandler:(void(^)(NSError *error))completionHandler;
 
+/**
+ Unsubscribes the device from change notifications for the specified tags.
+ @param tags An NSArray of tags
+ @param options (optional) an NSArray of the elements that you want to unsubscribe from. (CCHOptionBeacon, CCHOptionGeofence, CCHOptionVault, CCHOptionDevice)
+ @param completionHandler (optional) Is executed when the request completes.  If an error occurs, the NSError will be passed to the block.
+ */
+- (void)removeSubscriptionsForTags:(NSArray *)tags options:(NSArray *)options completionHandler:(void(^)(NSError *error))completionHandler;
 @end